refactor: migrate ViewModel tests to composeApp/commonTest (#372)
Move 14 ViewModel tests, 3 data/parsing tests, 13 fakes, and TestData
from app/src/test/ to composeApp/src/commonTest/ so they run on both
Android and Desktop JVM targets via :composeApp:desktopTest.

Conversions applied per file:
- org.junit.Test → kotlin.test.Test
- org.junit.Assert.* → kotlin.test.* (assertEquals, assertTrue, etc.)
- @Before → @BeforeTest, @After → @AfterTest
- MainDispatcherRule (JUnit4 rule) → setupMainDispatcher()/tearDownMainDispatcher()
- @Test(expected = ...) → assertFailsWith<>()
- System.currentTimeMillis() → fixed values (KMP-compatible)
- String(ByteArray, Charset) → ByteArray.decodeToString()

Fakes remain duplicated in app/src/test/fakes/ because the 9
Robolectric UI screen tests still reference them and app/ cannot
access composeApp/commonTest/ classes. This duplication is temporary
and will be resolved when the UI tests are migrated in a follow-up.

Test count preserved: 204 (composeApp/commonTest) + 104 (app) +
229 (shared) = 537 total.
